Arrumando o return de algumas classes de herois

+++ projeto/src/Acoes.java
@@ -2,7 +2,7 @@ public interface Acoes{
     abstract public boolean verificarVida ();
     abstract public void matar();
     abstract public void reanimar();
-    abstract public int[] executarMissao();
+    abstract public double[] executarMissao();
     abstract public boolean darXp(int xp);
     abstract public void evoluir(String atributo);
     abstract public void descansar();
