+++ projeto/src/ManaSabia.java
@@ -7,13 +7,13 @@ public class ManaSabia extends Herois{
     }
 
     @Override
-    public int[] executarMissao() {
+    public double[] executarMissao() {
         if (fadigaMental < 100){
             this.fadigaMental -= 25;
             return super.executarMissao();
         }
         else {
-            int[] erro = {6};
+            double[] erro = {6};
             return erro;
         }
     }
